add mixers to hyper_params_factory
