+++ pyriemann_qiskit/utils/__init__.py
@@ -7,6 +7,7 @@ from .docplex import (
     square_bin_mat_var,
     ClassicalOptimizer,
     NaiveQAOAOptimizer,
+    QAOACVOptimizer,
     set_global_optimizer,
     get_global_optimizer,
 )
@@ -36,6 +37,7 @@ __all__ = [
     "square_bin_mat_var",
     "ClassicalOptimizer",
     "NaiveQAOAOptimizer",
+    "QAOACVOptimizer",
     "set_global_optimizer",
     "get_global_optimizer",
     "distance",
